requirements.txt
alpaca-py
alpaca-trade-api
pandas
numpy
python-dotenv
requests
pytz
discord.py
pyzmq
pyarrow
